Removing dead code
